feat:게시판 관련 수정 및 SecurityCongfig 추가2

+++ Backend/doodle/src/main/java/com/cyworld/minihompy/dto/PostResponseDto.java
@@ -12,7 +12,7 @@ public class PostResponseDto {
     private String content;
     private String authorNickname;
     private LocalDateTime createdAt;
-    private Integer viewCount;
+    private Long viewCount;
 
     public PostResponseDto(Post post) {
         this.postId = post.getPostId();
